[Refactor] #643 - App.vue Header를 common/NavBar.vue로 분리
- Header 템플릿, 사용자 메뉴, 로그아웃 로직을 NavBar.vue로 이동
- App.vue 102줄 → 38줄로 축소, 레이아웃 껍데기만 유지
